+++ packages/wonder-blocks-dropdown/src/components/__tests__/select-opener.test.tsx
@@ -18,7 +18,6 @@ describe("SelectOpener", () => {
                     disabled={false}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}
@@ -43,7 +42,6 @@ describe("SelectOpener", () => {
                     disabled={false}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}
@@ -79,7 +77,6 @@ describe("SelectOpener", () => {
                     disabled={false}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}
@@ -123,7 +120,6 @@ describe("SelectOpener", () => {
                     disabled={true}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}
@@ -148,7 +144,6 @@ describe("SelectOpener", () => {
                     disabled={true}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}
@@ -184,7 +179,6 @@ describe("SelectOpener", () => {
                     disabled={true}
                     error={false}
                     isPlaceholder={false}
-                    light={false}
                     open={false}
                 >
                     {children}