+++ packages/wonder-blocks-dropdown/src/components/__tests__/single-select.test.tsx
@@ -839,7 +839,6 @@ describe("SingleSelect", () => {
             // Arrange
             const {userEvent} = doRender(
                 <SingleSelect
-                    light={true}
                     onChange={onChange}
                     isFilterable={true}
                     placeholder="Choose"