+++ packages/wonder-blocks-dropdown/src/components/multi-select.tsx
@@ -79,11 +79,6 @@ type DefaultProps = Readonly<{
      */
     error?: boolean;
     /**
-     * Whether to display the "light" version of this component instead, for
-     * use when the component is used on a dark background.
-     */
-    light?: boolean;
-    /**
      * The values of the items that are currently selected.
      */
     selectedValues?: Array<string>;
@@ -238,7 +233,6 @@ type Props = AriaProps &
 const MultiSelect = (props: Props) => {
     const {
         id,
-        light = false,
         opener,
         testId,
         alignment = "left",
@@ -578,7 +572,6 @@ const MultiSelect = (props: Props) => {
                             aria-label={ariaLabel}
                             aria-controls={dropdownId}
                             isPlaceholder={menuContent === noneSelected}
-                            light={light}
                             onOpenChanged={handleOpenChanged}
                             onBlur={onOpenerBlurValidation}
                             open={open}
@@ -625,7 +618,6 @@ const MultiSelect = (props: Props) => {
                         ...getShortcuts(numEnabledOptions),
                         ...filteredItems,
                     ]}
-                    light={light}
                     onOpenChanged={handleOpenChanged}
                     open={open}
                     opener={renderOpener(