+++ packages/wonder-blocks-dropdown/src/components/select-opener.tsx
@@ -3,11 +3,14 @@ import {StyleSheet} from "aphrodite";
 
 import {keys, type AriaProps} from "@khanacademy/wonder-blocks-core";
 
-import {mix} from "@khanacademy/wonder-blocks-tokens";
 import {addStyle} from "@khanacademy/wonder-blocks-core";
 import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
 import {LabelMedium} from "@khanacademy/wonder-blocks-typography";
-import * as tokens from "@khanacademy/wonder-blocks-tokens";
+import {
+    border,
+    semanticColor,
+    spacing,
+} from "@khanacademy/wonder-blocks-tokens";
 import caretDownIcon from "@phosphor-icons/core/bold/caret-down-bold.svg";
 import {DROPDOWN_ITEM_HEIGHT} from "../util/constants";
 import {OptionLabel} from "../util/types";
@@ -43,11 +46,6 @@ type SelectOpenerProps = AriaProps & {
      */
     ariaLabel?: string;
     /**
-     * Whether to display the "light" version of this component instead, for
-     * use when the item is used on a dark background.
-     */
-    light: boolean;
-    /**
      * Callback for when the SelectOpener is pressed.
      */
     onOpenChanged: (open: boolean) => unknown;
@@ -68,7 +66,6 @@ type SelectOpenerProps = AriaProps & {
 type DefaultProps = {
     disabled: SelectOpenerProps["disabled"];
     error: SelectOpenerProps["error"];
-    light: SelectOpenerProps["light"];
     isPlaceholder: SelectOpenerProps["isPlaceholder"];
 };
 
@@ -93,7 +90,6 @@ export default class SelectOpener extends React.Component<
     static defaultProps: DefaultProps = {
         disabled: false,
         error: false,
-        light: false,
         isPlaceholder: false,
     };
 
@@ -138,7 +134,6 @@ export default class SelectOpener extends React.Component<
             error,
             id,
             isPlaceholder,
-            light,
             open,
             testId,
             "aria-label": ariaLabel,
@@ -149,23 +144,17 @@ export default class SelectOpener extends React.Component<
             ...sharedProps
         } = this.props;
 
-        const stateStyles = _generateStyles(light, isPlaceholder, error);
+        const stateStyles = _generateStyles(isPlaceholder, error);
 
-        // The icon colors are kind of fickle. This is just logic
-        // based on the zeplin design.
-        const iconColor = light
-            ? disabled || error
-                ? "currentColor"
-                : tokens.color.white
-            : disabled
-            ? tokens.color.offBlack32
-            : tokens.color.offBlack64;
+        const iconColor = disabled
+            ? semanticColor.action.disabled.default
+            : semanticColor.icon.primary;
 
         const style = [
             styles.shared,
             stateStyles.default,
             disabled && stateStyles.disabled,
-            !disabled && this.state.pressed && stateStyles.pressed,
+            !disabled && this.state.pressed && stateStyles.press,
         ];
 
         return (
@@ -212,15 +201,15 @@ const styles = StyleSheet.create({
         display: "inline-flex",
         alignItems: "center",
         justifyContent: "space-between",
-        color: tokens.color.offBlack,
+        color: semanticColor.text.primary,
         height: DROPDOWN_ITEM_HEIGHT,
         // This asymmetry arises from the Icon on the right side, which has
         // extra padding built in. To have the component look more balanced,
         // we need to take off some paddingRight here.
-        paddingLeft: tokens.spacing.medium_16,
-        paddingRight: tokens.spacing.small_12,
+        paddingLeft: spacing.medium_16,
+        paddingRight: spacing.small_12,
         borderWidth: 0,
-        borderRadius: tokens.border.radius.medium_4,
+        borderRadius: border.radius.medium_4,
         borderStyle: "solid",
         outline: "none",
         textDecoration: "none",
@@ -232,7 +221,7 @@ const styles = StyleSheet.create({
     },
 
     text: {
-        marginRight: tokens.spacing.xSmall_8,
+        marginRight: spacing.xSmall_8,
         whiteSpace: "nowrap",
         userSelect: "none",
         overflow: "hidden",
@@ -244,134 +233,106 @@ const styles = StyleSheet.create({
     },
 });
 
-// These values are default padding (16 and 12) minus 1, because
-// changing the borderWidth to 2 messes up the button width
-// and causes it to move a couple pixels. This fixes that.
-const adjustedPaddingLeft = tokens.spacing.medium_16 - 1;
-const adjustedPaddingRight = tokens.spacing.small_12 - 1;
-
 const stateStyles: Record<string, any> = {};
 
-const _generateStyles = (
-    light: boolean,
-    placeholder: boolean,
-    error: boolean,
-) => {
+const _generateStyles = (placeholder: boolean, error: boolean) => {
     // "hash" the parameters
-    const styleKey = `${light}-${placeholder}-${error}`;
+    const styleKey = `${placeholder}-${error}`;
     if (stateStyles[styleKey]) {
         return stateStyles[styleKey];
     }
 
-    let newStyles: Record<string, any> = {};
-    if (light) {
-        const focusHoverStyling = {
-            borderColor: error ? tokens.color.red : tokens.color.white,
-            borderWidth: tokens.spacing.xxxxSmall_2,
-            paddingLeft: adjustedPaddingLeft,
-            paddingRight: adjustedPaddingRight,
-        };
-        const activePressedStyling = {
-            paddingLeft: adjustedPaddingLeft,
-            paddingRight: adjustedPaddingRight,
-            borderColor: error ? tokens.color.red : tokens.color.fadedBlue,
-            borderWidth: tokens.border.width.thin,
-            color: error
-                ? tokens.color.offBlack64
-                : placeholder
-                ? mix(tokens.color.white32, tokens.color.blue)
-                : tokens.color.fadedBlue,
-            backgroundColor: error
-                ? tokens.color.fadedRed
-                : tokens.color.activeBlue,
-        };
-        newStyles = {
-            default: {
-                background: error ? tokens.color.fadedRed8 : "transparent",
-                color: error
-                    ? tokens.color.offBlack64
-                    : placeholder
-                    ? tokens.color.white50
-                    : tokens.color.white,
-                borderColor: error ? tokens.color.red : tokens.color.white50,
-                borderWidth: tokens.border.width.hairline,
-                ":hover:not([aria-disabled=true])": focusHoverStyling,
-                // Allow hover styles on non-touch devices only. This prevents an
-                // issue with hover being sticky on touch devices (e.g. mobile).
-                ["@media not (hover: hover)"]: {
-                    ":hover:not([aria-disabled=true])": {
-                        borderColor: error
-                            ? tokens.color.red
-                            : tokens.color.white50,
-                        borderWidth: tokens.border.width.hairline,
-                        paddingLeft: tokens.spacing.medium_16,
-                        paddingRight: tokens.spacing.small_12,
-                    },
-                },
-                ":focus-visible:not([aria-disabled=true])": focusHoverStyling,
-                ":active:not([aria-disabled=true])": activePressedStyling,
-            },
-            disabled: {
-                background: "transparent",
-                borderColor: mix(tokens.color.white32, tokens.color.blue),
-                color: mix(tokens.color.white32, tokens.color.blue),
-                cursor: "not-allowed",
-                ":focus-visible": {
-                    boxShadow: `0 0 0 1px ${tokens.color.offBlack32}, 0 0 0 3px ${tokens.color.fadedBlue}`,
-                },
-            },
-            pressed: activePressedStyling,
-        };
-    } else {
-        const focusHoverStyling = {
-            borderColor: error ? tokens.color.red : tokens.color.blue,
-            borderWidth: tokens.border.width.thin,
-            paddingLeft: adjustedPaddingLeft,
-            paddingRight: adjustedPaddingRight,
-        };
-        const activePressedStyling = {
-            background: error ? tokens.color.fadedRed : tokens.color.fadedBlue,
-            borderColor: error ? tokens.color.red : tokens.color.activeBlue,
-            borderWidth: tokens.border.width.thin,
-            paddingLeft: adjustedPaddingLeft,
-            paddingRight: adjustedPaddingRight,
-        };
-        newStyles = {
-            default: {
-                background: error ? tokens.color.fadedRed8 : tokens.color.white,
-                borderColor: error ? tokens.color.red : tokens.color.offBlack50,
-                borderWidth: tokens.border.width.hairline,
-                color: placeholder
-                    ? tokens.color.offBlack64
-                    : tokens.color.offBlack,
-                ":hover:not([aria-disabled=true])": focusHoverStyling,
-                // Allow hover styles on non-touch devices only. This prevents an
-                // issue with hover being sticky on touch devices (e.g. mobile).
-                ["@media not (hover: hover)"]: {
-                    ":hover:not([aria-disabled=true])": {
-                        borderColor: error
-                            ? tokens.color.red
-                            : tokens.color.offBlack50,
-                        borderWidth: tokens.border.width.hairline,
-                        paddingLeft: tokens.spacing.medium_16,
-                        paddingRight: tokens.spacing.small_12,
-                    },
-                },
-                ":focus-visible:not([aria-disabled=true])": focusHoverStyling,
-                ":active:not([aria-disabled=true])": activePressedStyling,
-            },
-            disabled: {
-                background: tokens.color.offWhite,
-                borderColor: tokens.color.offBlack16,
-                color: tokens.color.offBlack64,
-                cursor: "not-allowed",
-                ":focus-visible": {
-                    boxShadow: `0 0 0 1px ${tokens.color.white}, 0 0 0 3px ${tokens.color.offBlack32}`,
+    // The different states that the component can be in.
+    const states = {
+        // Resting state
+        default: {
+            border: semanticColor.border.strong,
+            background: semanticColor.surface.primary,
+            foreground: semanticColor.text.primary,
+        },
+        disabled: {
+            border: semanticColor.border.primary,
+            background: semanticColor.action.disabled.secondary,
+            foreground: semanticColor.text.secondary,
+        },
+        // Form validation error state
+        error: {
+            border: semanticColor.status.critical.foreground,
+            background: semanticColor.status.critical.background,
+            foreground: semanticColor.text.primary,
+        },
+    };
+
+    // The color is based on the action color.
+    const actionType = error ? "destructive" : "progressive";
+    // NOTE: We are using the outlined action type for all the non-resting
+    // states as the opener is a bit different from a regular button in its
+    // resting/default state.
+    const action = semanticColor.action.outlined[actionType];
+
+    // TODO(WB-1856): Define global semantic outline tokens.
+    const sharedOutlineStyling = {
+        // Outline sits inside the border (inset)
+        outlineOffset: -border.width.thin,
+        outlineStyle: "solid",
+        outlineWidth: border.width.thin,
+    };
+
+    const focusHoverStyling = {
+        // TODO(WB-1856): Use `border.focus` when we define a global pattern for
+        // focus indicators.
+        outlineColor: action.hover.border,
+        ...sharedOutlineStyling,
+    };
+    const pressStyling = {
+        background: action.press.background,
+        color: placeholder
+            ? error
+                ? semanticColor.text.secondary
+                : semanticColor.action.outlined.progressive.press.foreground
+            : semanticColor.text.primary,
+        outlineColor: action.press.border,
+        ...sharedOutlineStyling,
+    };
+
+    const currentState = error ? states.error : states.default;
+
+    const newStyles = {
+        default: {
+            background: currentState.background,
+            borderColor: currentState.border,
+            borderWidth: border.width.hairline,
+            color: placeholder
+                ? semanticColor.text.secondary
+                : currentState.foreground,
+            ":hover:not([aria-disabled=true])": focusHoverStyling,
+            // Allow hover styles on non-touch devices only. This prevents an
+            // issue with hover being sticky on touch devices (e.g. mobile).
+            ["@media not (hover: hover)"]: {
+                ":hover:not([aria-disabled=true])": {
+                    borderColor: currentState.border,
+                    borderWidth: border.width.hairline,
+                    paddingLeft: spacing.medium_16,
+                    paddingRight: spacing.small_12,
                 },
             },
-            pressed: activePressedStyling,
-        };
-    }
+            ":focus-visible:not([aria-disabled=true])": focusHoverStyling,
+            ":active:not([aria-disabled=true])": pressStyling,
+        },
+        disabled: {
+            background: states.disabled.background,
+            borderColor: states.disabled.border,
+            color: states.disabled.foreground,
+            cursor: "not-allowed",
+            ":focus-visible": {
+                // TODO(WB-1856): Use `border.focus` when we define a global
+                // pattern for focus indicators.
+                outlineColor: semanticColor.action.disabled.default,
+                ...sharedOutlineStyling,
+            },
+        },
+        press: pressStyling,
+    };
 
     stateStyles[styleKey] = StyleSheet.create(newStyles);
     return stateStyles[styleKey];