+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.0.6
+
+### Patch Changes
+
+-   Updated dependencies [0de25cd8]
+    -   @khanacademy/wonder-blocks-tokens@4.0.0
+
 ## 5.0.5
 
 ### Patch Changes